+++ changelog.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setup GitHub workflow.
 - Setup default plugin structure.
 - Setup first version of metadata.
+- Setup first version of parser.